Le pop-up pour l'actualisation de la page.
