+++ src/App.css
@@ -40,3 +40,39 @@
 /*.read-the-docs {*/
 /*  color: #888;*/
 /*}*/
+.modal-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.modal-content {
+    background: white;
+    padding: 20px;
+    border-radius: 10px;
+    text-align: center;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.3);
+}
+
+.buttonModal {
+    margin: 10px;
+    padding: 10px 20px;
+    border: none;
+    cursor: pointer;
+}
+
+.buttonModal:first-of-type {
+    background: red;
+    color: white;
+}
+
+.buttonModal:last-of-type {
+    background: green;
+    color: white;
+}
