+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css'
 import {Outlet} from "react-router";
-import {UnloadLogout} from "./utils/UnloadLogout.ts";
+import {UnloadLogout} from "./utils/UnloadLogout.tsx";
 
 export const App = () => {
     UnloadLogout()
