+++ src/services_REST/serveur/Disconnect.ts
@@ -1,12 +1,12 @@
-import axios from "axios"
-import parametres from "../../../public/parameters.json"
+import parameters from "../../../public/parameters.json"
+import axios from "axios";
 
-const URL_SERVEUR = parametres.URL_SERVER
-const URL_LOGOUT = parametres.URL_LOGOUT
+const URL_SERVER = parameters.URL_SERVER
+const URL_LOGOUT = parameters.URL_LOGOUT
 
 export const Disconnect = async () => {
     try {
-        await axios.get(`${URL_SERVEUR}${URL_LOGOUT}`, {withCredentials: true})
+        await axios.get(`${URL_SERVER}${URL_LOGOUT}`, {withCredentials: true})
     } catch (error) {
         console.error(error)
     }
